+++ aeryn_core/audit_trail/__init__.py
@@ -1,0 +1,2 @@
+from .trail import AuditTrail, audit_trail
+__all__ = ['AuditTrail', 'audit_trail']