+++ aeryn_core/job_queue/__init__.py
@@ -1,0 +1,2 @@
+from .queue import JobQueue, job_queue
+__all__ = ['JobQueue', 'job_queue']